Refactor regular expression creation into a function
The same code existed in two places to generate more or less identical
regular expressions for the skipped and moved checks.  This commit
refactors that code into a function and uses that function in the two
places where the regular expressions were generated.  This should make
maintenance easier if the regular expressions ever need to change.

Signed-off-by: Jason Stephenson <jason@sigio.com>